compressors.lib: typo in comments
Everywhere in the comments, *_FFFB_* (Feed-Forward FeedBack) is used but in the actual dsp code it is *_FBFF_*. This leads to Faust IDE incorrectly parsing the doc and suggesting an unknown bit of code for the compiler.
It would have been easier to only replace *_FBFF_* by *_FFFB_* (2 occurences) but this would have broken any project based on those definitions.